requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
click==8.2.1
pre-commit==4.2.0
pydantic==2.11.5
pytest==8.3.5
python-dotenv==1.1.0
black==25.1.0
dash==2.16.1
openpyxl==3.1.5
pandera==0.24.0